+++ packages/site/.eslintrc.cjs
@@ -8,7 +8,7 @@ module.exports = {
     "import/no-unresolved": [
       "error",
       {
-        ignore: ["^@atlantis/docs"],
+        ignore: ["^@atlantis/docs", "^@atlantis/packages"],
       },
     ],
   },