+++ packages/site/src/components/ComponentLinks.tsx
@@ -1,5 +1,5 @@
 import { Box, Content, Heading, Icon, Link } from "@jobber/components";
-import { MouseEvent, useEffect, useState } from "react";
+import { AnchorLinks } from "./AnchorLinks";
 import { ContentExportLinks } from "../types/content";
 import { useAtlantisSite } from "../providers/AtlantisSiteProvider";
 
@@ -23,116 +23,71 @@ export const ComponentLinks = ({
   readonly webEnabled: boolean;
   readonly mobileEnabled: boolean;
 }) => {
-  const [hlinks, setHlinks] = useState<Element[] | null>(null);
-
-  useEffect(() => {
-    const hdd = document.querySelectorAll("[data-heading-link]");
-
-    if (hdd.length > 0) {
-      setHlinks(Array.from(hdd));
-    }
-  }, []);
-
-  const click = (e: MouseEvent) => {
-    e.preventDefault();
-    const id = e.currentTarget?.getAttribute("href")?.replace("#", "");
-
-    if (id) {
-      goToDesign();
-      setTimeout(() => {
-        const element = document.getElementById(id);
-
-        if (element) {
-          element.scrollIntoView({ behavior: "smooth" });
-        }
-      }, 100);
-    }
-  };
   const { isMinimal } = useAtlantisSite();
   if (isMinimal) return null;
 
   return (
-    <aside
-      style={{
-        width: "200px",
-        padding: "36px var(--space-base)",
-        display: "flex",
-        flexDirection: "column",
-        flexShrink: "0",
-        boxSizing: "border-box",
-      }}
-    >
-      <Content spacing={"larger"}>
+    <Content spacing={"larger"}>
+      <AnchorLinks
+        id="design"
+        header="Design"
+        additionalOnClickAction={goToDesign}
+      />
+      {webEnabled && (
         <Content>
           <Heading level={6} element="h3">
-            Design
+            Web
           </Heading>
           <Content spacing="small">
-            {hlinks?.map((link, index) => (
-              <Box key={index}>
-                <a onClick={click} href={`#${link.id}`}>
-                  {link.textContent}
-                </a>
-              </Box>
-            ))}
-          </Content>
-        </Content>
-        {webEnabled && (
-          <Content>
-            <Heading level={6} element="h3">
-              Web
-            </Heading>
-            <Content spacing="small">
-              <Box>
-                <a onClick={() => goToUsage("web")} href={`#`}>
-                  Usage
-                </a>
-              </Box>
-              <Box>
-                <a onClick={() => goToProps("web")} href={`#`}>
-                  Props
-                </a>
-              </Box>
-            </Content>
-          </Content>
-        )}
-        {mobileEnabled && (
-          <Content>
-            <Heading level={6} element="h3">
-              Mobile
-            </Heading>
-            <Content spacing="small">
-              <Box>
-                <a onClick={() => goToUsage("mobile")} href={`#`}>
-                  Usage
-                </a>
-              </Box>
-              <Box>
-                <a onClick={() => goToProps("mobile")} href={`#`}>
-                  Props
-                </a>
-              </Box>
-            </Content>
-          </Content>
-        )}
-        <Content>
-          <Heading level={6} element="h3">
-            Links
-          </Heading>
-          <Content spacing="smaller">
-            <Box direction="row" gap="smaller" alignItems="center">
-              <Icon size="small" color="interactive" name="link" />
-              {links?.map((link, index) => (
-                <div key={index} data-storybook-link>
-                  <Link key={index} url={link.url} external>
-                    {link.label}
-                  </Link>
-                </div>
-              ))}
+            <Box>
+              <a onClick={() => goToUsage("web")} href={`#`}>
+                Usage
+              </a>
+            </Box>
+            <Box>
+              <a onClick={() => goToProps("web")} href={`#`}>
+                Props
+              </a>
             </Box>
           </Content>
         </Content>
+      )}
+      {mobileEnabled && (
+        <Content>
+          <Heading level={6} element="h3">
+            Mobile
+          </Heading>
+          <Content spacing="small">
+            <Box>
+              <a onClick={() => goToUsage("mobile")} href={`#`}>
+                Usage
+              </a>
+            </Box>
+            <Box>
+              <a onClick={() => goToProps("mobile")} href={`#`}>
+                Props
+              </a>
+            </Box>
+          </Content>
+        </Content>
+      )}
+      <Content>
+        <Heading level={6} element="h3">
+          Links
+        </Heading>
+        <Content spacing="smaller">
+          <Box direction="row" gap="smaller" alignItems="center">
+            <Icon size="small" color="interactive" name="link" />
+            {links?.map((link, index) => (
+              <div key={index} data-storybook-link>
+                <Link key={index} url={link.url} external>
+                  {link.label}
+                </Link>
+              </div>
+            ))}
+          </Box>
+        </Content>
       </Content>
-    </aside>
+    </Content>
   );
 };