+++ packages/site/src/components/ContentLoader.tsx
@@ -1,5 +1,5 @@
 import { useLocation, useParams } from "react-router";
-import { ContentView } from "../pages/ContentView";
+import { ContentView } from "../layout/ContentView";
 import { contentMap } from "../maps";
 
 /**
@@ -19,8 +19,14 @@ export const ContentLoader = () => {
     case location.pathname.startsWith("/changelog"):
       type = "changelog";
       break;
+    case location.pathname.startsWith("/hooks"):
+      type = "hooks";
+      break;
     case location.pathname.startsWith("/guides"):
       type = "guides";
+      break;
+    case location.pathname.startsWith("/packages"):
+      type = "packages";
       break;
     default:
       type = "content";
@@ -30,8 +36,8 @@ export const ContentLoader = () => {
 
   return (
     <ContentView
-      intro={content.intro}
       title={content.title}
+      key={`${type}-${name}`}
       content={content.content}
     />
   );