+++ packages/site/src/pages/HomePage.tsx
@@ -5,7 +5,7 @@ export const HomePage = () => {
     <PageBlock
       structure={{
         header: {
-          title: "Atlantis",
+          title: "Home",
           body: "Jobber's toolkit for building consumer-grade experiences",
           ctaLabel: "Get Started",
           to: "/design/welcome-guide",
@@ -17,17 +17,46 @@ export const HomePage = () => {
               title: "Components",
               to: "/components",
               imageURL: "/Components.png",
+              sections: ["Packages"],
             },
-            { title: "Design", to: "/design", imageURL: "/Design.png" },
-            { title: "Content", to: "/content", imageURL: "/Placeholder.png" },
-            { title: "Guides", to: "/guides", imageURL: "/Placeholder.png" },
+            {
+              title: "Design",
+              to: "/design",
+              imageURL: "/Design.png",
+              sections: ["Packages"],
+            },
+            {
+              title: "Content",
+              to: "/content",
+              imageURL: "/Placeholder.png",
+              sections: ["Resources"],
+            },
+            {
+              title: "Hooks",
+              to: "/hooks",
+              imageURL: "/Hooks.png",
+              sections: ["Packages"],
+            },
+            {
+              title: "Guides",
+              to: "/guides",
+              imageURL: "/Placeholder.png",
+              sections: ["Resources"],
+            },
+            {
+              title: "Packages",
+              to: "/packages",
+              imageURL: "/Placeholder.png",
+            },
             {
               title: "Changelog",
               to: "/changelog",
               imageURL: "/Placeholder.png",
+              sections: ["Changelog"],
             },
           ],
         },
+        useCategories: true,
       }}
     />
   );